chore(jans-cedarling): refactor test/utils module, split functions to different folders
